requirements.txt
numpy
pandas
scikit-learn
python-chess